+++ lib/cn/vblock_reader.h
@@ -110,11 +110,11 @@ vbr_madvise_async(
 void
 vbr_readahead(
     struct vblock_desc *     vbd,
-    u32                      off,
-    u32                      vlen,
-    u32                      ra_flags,
-    u32                      ra_len,
-    u32                      ra_histc,
+    uint32_t                 off,
+    size_t                   vlen,
+    uint32_t                 ra_flags,
+    size_t                   ra_len,
+    uint32_t                 ra_histc,
     struct ra_hist *         ra_histv,
     struct workqueue_struct *wq);
